RECS.Compiler/BytecodeGenerator.cs
using System;
using System.Collections.Generic;
using System.Text;
using System.Text.Json;
using RECS.Core;

namespace RECS.Compiler
{
    public enum Opcode
    {
        NOP = 0,
        JUMP_IF_FALSE = 1,
        JUMP_IF_TRUE = 2,
        EXECUTE_ACTION = 3,
        // Define other opcodes as needed
    }

    public class Instruction
    {
        public Opcode Opcode { get; set; }
        public List<byte> Operands { get; set; } = new List<byte>();

        public int Size => 1 + Operands.Count;  // 1 byte for opcode + size of operands
    }

    public class BytecodeGenerator
    {
        public List<Instruction> Generate(Ruleset ruleset)
        {
            var instructions = new List<Instruction>();

            foreach (var rule in ruleset.Rules)
            {
                instructions.AddRange(GenerateInstructionsForRule(rule));
            }

            return instructions;
        }

        private List<Instruction> GenerateInstructionsForRule(Rule rule)
        {
            var instructions = new List<Instruction>();

            // Generate condition checking instructions
            foreach (var condition in rule.Conditions.All ?? new List<ConditionOrGroup>())
            {
                instructions.Add(GenerateConditionInstruction(condition, false));
            }

            foreach (var condition in rule.Conditions.Any ?? new List<ConditionOrGroup>())
            {
                instructions.Add(GenerateConditionInstruction(condition, true));
            }

            // Generate action execution instructions
            foreach (var action in rule.Actions)
            {
                instructions.Add(GenerateActionInstruction(action));
            }

            return instructions;
        }

        private Instruction GenerateConditionInstruction(ConditionOrGroup condition, bool isAnyCondition)
        {
            // Here we can implement logic for converting conditions into instructions
            var instruction = new Instruction
            {
                Opcode = isAnyCondition ? Opcode.JUMP_IF_TRUE : Opcode.JUMP_IF_FALSE
            };

            // Serialize the condition fact, operator, and value into bytes
            var factBytes = Encoding.UTF8.GetBytes(condition.Fact);
            instruction.Operands.AddRange(factBytes);

            // For simplicity, we can assume operator and value are serialized similarly, but this could be extended
            var operatorBytes = Encoding.UTF8.GetBytes(condition.Operator);
            instruction.Operands.AddRange(operatorBytes);

            var valueBytes = Encoding.UTF8.GetBytes(condition.Value.ToString());
            instruction.Operands.AddRange(valueBytes);

            return instruction;
        }

        private Instruction GenerateActionInstruction(RuleAction action)
        {
            var instruction = new Instruction
            {
                Opcode = Opcode.EXECUTE_ACTION
            };

            // Serialize the action type and target into bytes
            var typeBytes = Encoding.UTF8.GetBytes(action.Type);
            instruction.Operands.AddRange(typeBytes);

            var targetBytes = Encoding.UTF8.GetBytes(action.Target);
            instruction.Operands.AddRange(targetBytes);

            // The value can also be serialized, depending on the action type
            var valueBytes = Encoding.UTF8.GetBytes(action.Value.ToString());
            instruction.Operands.AddRange(valueBytes);

            return instruction;
        }
    }
}
